Убирает повторы индексов
